+++ DevelopmentApp/android/app/build.gradle
@@ -76,10 +76,10 @@ android {
     buildToolsVersion rootProject.ext.buildToolsVersion
     compileSdk rootProject.ext.compileSdkVersion
 
-    namespace "com.fluiddocs"
+    namespace "com.platformbuilders.fluid"
 
     defaultConfig {
-        applicationId "com.fluiddocs"
+        applicationId "com.platformbuilders.fluid"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
