+++ DevelopmentApp/app.json
@@ -1,4 +1,4 @@
 {
-  "name": "fluidDocs",
-  "displayName": "fluidDocs"
+  "name": "Fluid",
+  "displayName": "Fluid"
 }
