+++ DevelopmentApp/ios/Podfile
@@ -25,7 +25,7 @@ if linkage != nil
   use_frameworks! :linkage => linkage.to_sym
 end
 
-target 'fluidDocs' do
+target 'Fluid' do
   config = use_native_modules!
 
   use_react_native!(
@@ -39,7 +39,7 @@ target 'fluidDocs' do
     :app_path => "#{Pod::Config.instance.installation_root}/.."
   )
 
-  target 'fluidDocsTests' do
+  target 'FluidTests' do
     inherit! :complete
     # Pods for testing
   end
