+++ DevelopmentApp/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "fluid-docs",
+  "name": "fluid",
   "version": "0.0.1",
   "private": true,
   "scripts": {
